server/requirements.txt
asgiref==3.5.2
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==38.0.3
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
entrypoints==0.3
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.8.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numpy==1.21.3
oauthlib==3.2.2
openpyxl==3.0.9
pandas==1.3.4
parso==0.8.2
passlib==1.7.4
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.21
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2021.3
pyzmq==22.3.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tornado==6.1
traitlets==5.1.0
typing_extensions==4.4.0
urllib3==1.26.9
wcwidth==0.2.5
